% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csi.R
\name{csi}
\alias{csi}
\title{critical success index (CSI) or threat score (TS)}
\usage{
csi(A, B)
}
\arguments{
\item{\code{A}, }{a matrix(True parameter).}

\item{\code{B}, }{a matrix(Estimation).}
}
\value{
critical success index (CSI).
}
\description{
critical success index (CSI) or threat score (TS)
}
