% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gram}
\alias{gram}
\title{Weighted Gram matrix}
\usage{
gram(X, w)
}
\arguments{
\item{\code{X}, }{a \eqn{n \times p} matrix.}

\item{\code{w}, }{a \eqn{n} dimensional weight vector.}
}
\value{
a \eqn{p \times p} matrix.
}
\description{
Calculate the weighted Gram matrix:\eqn{X^\top diag(w)X}.
}
