% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{inf_vec}
\alias{inf_vec}
\title{Proximal projection of a \eqn{\ell_\infty} penalty.}
\usage{
inf_vec(x, lambda = 0)
}
\arguments{
\item{\code{x}, }{a vector.}

\item{\code{lambda}, }{a scalar.}
}
\value{
a vector after proximal projection.
}
\description{
Proximal projection solution of a vector which is the solution:
\deqn{\argmin_y ||y-x||_2^2/2+\lambda ||y||_\infty.}
}
