% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{l1inf_mat}
\alias{l1inf_mat}
\title{Proximal projection of a Matrix with a hybrid \eqn{\ell_1/\ell_\infty} penalty.}
\usage{
l1inf_mat(X, lambda = 0, k = 1L)
}
\arguments{
\item{\code{X}}{a Matrix}

\item{\code{lambda}}{a scalar}

\item{\code{k}}{Not penalized the first \code{k} columns. Default is 1.}
}
\value{
A Matrix.
}
\description{
Proximal projection solution of a Matrix where each column is the solution:
\deqn{\argmin_y ||y-x||_2^2/2+\lambda  \max\{|y_1|, \sum_{i=2}^p |y_i|\}.}
}
