% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{l2_vec}
\alias{l2_vec}
\title{Proximal projection of a Vector with  \eqn{\ell_2} penalty (Group LASSO).}
\usage{
l2_vec(x, lambda = 0)
}
\arguments{
\item{\code{x}}{a vector}

\item{\code{lambda}}{a scalar}
}
\value{
A vector which is proximal projection solution.
}
\description{
Proximal projection solution of a vector which is the solution:
\deqn{\argmin_y ||y-x||_2^2/2+\lambda ||y||_2.}
}
