% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longvec2mat.R
\name{longvec2mat}
\alias{longvec2mat}
\title{Transform a long vector into a matrix}
\usage{
longvec2mat(aa)
}
\arguments{
\item{\code{aa}, }{a \eqn{p(p+3)/2} dimensional vector.}
}
\value{
a \eqn{(p+1)\times (p+1)} sparse matrix.
}
\description{
Transform a long vector into a matrix
}
