% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure.R
\name{measure}
\alias{measure}
\title{Transform a long vector into a matrix}
\usage{
measure(A, B)
}
\arguments{
\item{\code{A}, }{a matrix(True parameter).}

\item{\code{B}, }{a matrix(Estimation).}
}
\value{
a 4 dimensional vector with nnzero of \code{B}, Sensitivity, Specificity and also the Frobenious norm of A-B.
}
\description{
Transform a long vector into a matrix
}
