% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{nuclear_mat}
\alias{nuclear_mat}
\title{Proximal projection of a nuclear norm penalty.}
\usage{
nuclear_mat(X, lambda = 0)
}
\arguments{
\item{\code{X}, }{a matrix.}

\item{\code{lambda}, }{a scalar.}
}
\value{
a matrix.
}
\description{
Proximal projection solution of a matrix with a nuclear norm penalty: 
\deqn{\argmin_Y ||Y-X||_2^2/2+\lambda \|Y\|_*.}
}
