% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{qr1}
\alias{qr1}
\title{Quadratic regression with \eqn{\ell_1} penalty}
\usage{
qr1(
  X,
  Y,
  lambda,
  err_abs = 10^(-4),
  err_rel = 10^(-3),
  maxIter = 200L,
  rho = 1,
  rho_vary = 0L
)
}
\arguments{
\item{\code{X}, }{a \eqn{n*p} input data matrix.}

\item{\code{Y}, }{a \eqn{n} dimensional response vector.}

\item{\code{lambda}, }{user supplied tuning parameter.}

\item{\code{err_abs}, }{the absolute tolerance precision used to stop the convergence of ADMM.}

\item{\code{err_rel}, }{the relative tolerance precision used to stop the convergence of ADMM.}

\item{\code{maxIter}, }{Maximum number of iterations. Default is 1000.}

\item{\code{rho}, }{initial step parameter for ADMM.}

\item{\code{rho_vary}, }{whether varying penalty parameter \eqn{\rho} for ADMM. Default is 0 (no varying).}
}
\value{
A list with components
\item{Omega}{a list of sparse p*p matrices corresponding to lambda.}
\item{lambda}{the used lambda for the solution path.}
\item{niter}{the number of iterations for each element of lambda.}
}
\description{
ADMM algorithm for high dimensional Quadratic regression with a \eqn{\ell_1} norm penalty:
\deqn{\argmin_Y \frac{1}{2n}\sum_{i=1}^n (Y_i-X_i^\top \Omega X_i)^2+\lambda \|\Omega\|_1.}
}
