% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{qr1_rank}
\alias{qr1_rank}
\title{Quadratic regression with nuclear norm penalty}
\usage{
qr1_rank(
  X,
  Y,
  lambda,
  err_abs = 10^(-4),
  err_rel = 10^(-3),
  maxIter = 200L,
  rho = 1
)
}
\arguments{
\item{\code{X}, }{a \eqn{n*p} input data matrix.}

\item{\code{Y}, }{a \eqn{n} response vector.}

\item{\code{lambda}, }{user supplied tuning parameter;}

\item{\code{err_abs}, }{\code{err_rel}   the precision used to stop the convergence of ADMM.}

\item{\code{maxIter}, }{Maximum number of iterations. Default is 1000.}

\item{\code{rho}, }{initial step parameter for ADMM.}
}
\value{
A list with components
\item{Omega}{a list of sparse p*p matrices corresponding to lambda.}
\item{lambda}{the used lambda for the solution path.}
\item{niter}{the number of iterations for each element of lambda.}
}
\description{
ADMM algorithm for high dimensional Quadratic regression with a \eqn{\ell_1} norm penalty:
\deqn{\argmin_Y \frac{1}{2n}\sum_{i=1}^n (Y_i-X_i^\top \Omega X_i)^2+\lambda \|\Omega\|_*.}
}
