% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{qr3_rank}
\alias{qr3_rank}
\title{Quadratic regression with \eqn{\ell_1} penalty and nuclear norm penalty (Sparse and Low rank)}
\usage{
qr3_rank(
  X,
  Y,
  lambda1,
  lambda2,
  err_abs = 10^(-4),
  err_rel = 10^(-3),
  maxIter = 200L,
  rho = 5
)
}
\arguments{
\item{\code{X}, }{a \eqn{n\times p} data matrix.}

\item{\code{Y}, }{a \eqn{n} dimensional response vector.}

\item{\code{lambda1}, }{user supplied tuning parameters.}

\item{\code{lambda2}, }{user supplied tuning parameters.}

\item{\code{err_abs}, }{\code{err_rel},   the precision used to stop the convergence of ADMM.}

\item{\code{maxIter}, }{maximum number of iterations. Default is 200.}

\item{\code{rho}, }{initial step parameter for ADMM.}
}
\value{
A list with components
\item{Omega}{a list of sparse \eqn{p \times p} matrices corresponding to lambda.}
\item{lambda1}{the used lambda1 for the solution path.}
\item{lambda2}{the used lambda2 for the solution path.}
\item{niter}{the number of iterations for each element of (lambda1,lambda2).}
}
\description{
ADMM algorithm for high dimensional Quadratic regression with \eqn{\ell_1} and nuclear norm penalty:
\deqn{\argmin_Y \frac{1}{2n}\sum_{i=1}^n (Y_i-X_i^\top \Omega X_i)^2+\lambda_1 \|\Omega\|_1+\lambda_2 \|\Omega\|_*.}
}
