% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{qrow}
\alias{qrow}
\title{Quadratic form for each sample}
\usage{
qrow(X, W)
}
\arguments{
\item{\code{X}, }{a \eqn{n \times p} matrix.}

\item{\code{W}, }{a \eqn{p \times p} weight matrix.}
}
\value{
a \eqn{n} dimensional vector.
}
\description{
Calculate the quadratic form for each sample with a Weight matrix \code{W}: 
\deqn{ diag(X W X^\top)=(X_1^\top W X_1,\cdots,X_n^\top W X_n)^\top.}
}
