% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ridge_qr}
\alias{ridge_qr}
\title{Quadratic regression with squared \eqn{\ell_2} penalty (Ridge regression)}
\usage{
ridge_qr(X, Y, lambda)
}
\arguments{
\item{\code{X}}{a \eqn{n*p} input data matrix.}

\item{\code{Y}}{a \eqn{n} response vector.}

\item{\code{lambda}}{user supplied tuning parameter;}
}
\value{
A list with components
\item{Omega}{a list of sparse p*p matrices corresponding to lambda.}
\item{lambda}{the used lambda.}
}
\description{
Algorithm for high dimensional Quadratic regression with squared \eqn{\ell_2} penalty
}
