% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{soft}
\alias{soft}
\title{Element-wise Soft Thresholding Function for a Matrix}
\usage{
soft(A, a)
}
\arguments{
\item{\code{A}}{Matrix}

\item{\code{a}}{scalar}
}
\value{
Matrix after threholding
}
\description{
Soft thresholding function for Matrix which is the solution of the  
\deqn{\argmin_X ||X-A||_2^2/2+a ||X||_1.}
}
