% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{soft_vec}
\alias{soft_vec}
\title{Element-wise Soft Thresholding Function for Vector}
\usage{
soft_vec(y, lambda)
}
\arguments{
\item{\code{y}}{a vector}

\item{\code{lambda}}{a scalar}
}
\value{
Vector after threholding
}
\description{
Soft thresholding function for a vector which is the solution of the  
\deqn{\argmin_x ||y-x||_2^2/2+\lambda ||x||_1.}
}
