% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vec2inter.R
\name{vec2inter}
\alias{vec2inter}
\title{Calculate a p2 interaction vector for a vector}
\usage{
vec2inter(a)
}
\arguments{
\item{\code{a}, }{a \eqn{p} dimensional vector.}
}
\value{
a \eqn{p^2} dimensional vector
}
\description{
Calculate a p2 interaction vector for a vector
}
