% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vec2longvec.R
\name{vec2longvec}
\alias{vec2longvec}
\title{Calculate a interaction vector without duplication for a vector}
\usage{
vec2longvec(a)
}
\arguments{
\item{\code{a}, }{a \eqn{p} dimensional vector.}
}
\value{
a \eqn{p(p+3)/2} dimensional vector
}
\description{
Calculate a interaction vector without duplication for a vector
}
