% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gram}
\alias{gram}
\title{Weighted Gram Matrix}
\usage{
gram(X, w)
}
\arguments{
\item{\code{X}}{a \eqn{n \times p} Matrix}

\item{\code{w}}{a \eqn{n} dimensional weight vector}
}
\value{
Weighted Gram Matrix
}
\description{
\code{gram} computes the weighted Gram matrix: 
\deqn{ X^\top diag(w) X.}
}
