% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{l2_mat}
\alias{l2_mat}
\title{Proximal projection of a \eqn{\ell_2} penalty for a Matrix}
\usage{
l2_mat(X, lambda = 0, k = 1L)
}
\arguments{
\item{\code{X}, }{a matrix.}

\item{\code{lambda}, }{a scalar.}

\item{\code{k}, }{not penalized the first \code{k} columns. Default is 1.}
}
\value{
A Matrix.
}
\description{
Proximal projection solution of a Matrix where each column is the solution:
\deqn{\argmin_y ||y-x||_2^2/2+\lambda ||y||_2.}
}
