% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{qr1}
\alias{qr1}
\title{Quadratic regression with \eqn{\ell_1} penalty}
\usage{
qr1(
  X,
  Y,
  lambda,
  err_abs = 10^(-4),
  err_rel = 10^(-3),
  maxIter = 200L,
  rho = 1
)
}
\arguments{
\item{\code{X}}{a \eqn{n*p} input data matrix.}

\item{\code{Y}}{a \eqn{n} response vector.}

\item{\code{lambda}, }{user supplied tuning parameter;}

\item{\code{err_abs}, }{\code{err_rel}   the precision used to stop the convergence of ADMM.}

\item{\code{maxIter}, }{Maximum number of iterations. Default is 1000.}

\item{\code{rho}, }{initial step parameter for ADMM.}
}
\value{
A list with components
\item{Omega}{a list of sparse p*p matrices corresponding to lambda.}
\item{lambda}{the used lambda for the solution path.}
\item{niter}{the number of iterations for each element of lambda.}
}
\description{
ADMM algorithm for high dimensional Quadratic regression with \eqn{\ell_1} penalty
}
