% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{qr2}
\alias{qr2}
\title{Quadratic regression with squared \eqn{\ell_2} penalty (Ridge regression)}
\usage{
qr2(X, Y, lambda)
}
\arguments{
\item{\code{X}, }{a \eqn{n \times p} data matrix.}

\item{\code{Y}, }{a \eqn{n} dimensional response vector.}

\item{\code{lambda}, }{user supplied tuning parameter;}
}
\value{
A list with components
\item{Omega}{a list of sparse \eqn{p \times p} matrices corresponding to lambda.}
\item{lambda}{the used lambda.}
}
\description{
Algorithm for High dimensional Quadratic Regression(HiQR) with squared \eqn{\ell_2} penalty:
\deqn{\argmin_Y \frac{1}{2n}\sum_{i=1}^n (Y_i-X_i^\top \Omega X_i)^2+\lambda \|\Omega\|_2^2.}
}
