% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{qrow}
\alias{qrow}
\title{Quadratic of Each Row}
\usage{
qrow(X, W)
}
\arguments{
\item{\code{X}}{a \eqn{n*p} Matrix}

\item{\code{W}}{a \eqn{p*p} weight Matrix}
}
\value{
Quadratic of Each Row
}
\description{
\code{qrow} computes the quadratic of Each Row : 
\deqn{ \diag(X W X^\top)=(X_1^\top W X_1,\cdots,X_n^\top W X_n)^\top}
}
