% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{soft}
\alias{soft}
\title{Element-wise soft-thresholding function for a matrix}
\usage{
soft(X, lambda = 0, k = 1L)
}
\arguments{
\item{\code{X}, }{a matrix.}

\item{\code{lambda}, }{a scalar.}

\item{\code{k}, }{not penalized the first element. Default is \code{1}(TRUE).}
}
\value{
a matrix after thresholding.
}
\description{
Soft-thresholding function for a matrix which is the solution of
\deqn{\argmin_Y ||Y-X||_2^2/2+\lambda ||Y||_1.}
}
